Split all basic tests into own package
